move tiptracker and wellgrid
