+++ api/src/opentrons/protocols/context/labware.py
@@ -5,9 +5,9 @@ from typing import List, Dict
 
 from opentrons.protocols.geometry.deck_item import DeckItem
 from opentrons.protocols.geometry.labware_geometry import AbstractLabwareGeometry
-from opentrons.protocols.geometry.tip_tracker import TipTracker
+from opentrons.protocols.api_support.tip_tracker import TipTracker
 from opentrons.protocols.context.well import WellImplementation
-from opentrons.protocols.geometry.well_grid import WellGrid
+from opentrons.protocols.api_support.well_grid import WellGrid
 from opentrons.types import Point
 from opentrons_shared_data.labware.dev_types import (
     LabwareParameters, LabwareDefinition)
